Avoid infinite recursion in WAYWindowDelegateProxyDelegate
Imagine that NSWindow implements a method -foo as follows:

```
- (id)foo {
	return [self.delegate foo];
}
```

Without any safeguards, the WAYWindowDelegateProxy would receive the invocation for -foo and attempt to forward it to its firstDelegate... which is the window itself. This leads to infinite recursion. This actually happens in OS X 10.10: -[NSWindow validRequestorForSendType:returnType:] forwards the call to its delegate if the delegate claims to be an instance of NSResponder, which the proxy usually does. There may well be other similar cases.

This change implements safeguard against this situation by keeping a stack of selectors in the process of being forwarded. If a recursive call is detected, the recursive one is aborted, which will have the effect of returning nil. This is about as safe as we can make things in a generic proxy like this.